Improve type safety (I hope) Make ESLint stricter Add prettier

+++ src/commands/randomisecolour.ts
@@ -6,18 +6,18 @@ import ClientPlus from "../classes/ClientPlus";
 export default {
     data: new SlashCommandBuilder()
         .setName("randomisecolour")
-        .setDescription("Toggles randomising your name colour when you post a message (may get overridden by other roles)"),
+        .setDescription(
+            "Toggles randomising your name colour when you post a message (may get overridden by other roles)"
+        ),
     async execute(interaction: CommandInteraction) {
         const client = interaction.client as ClientPlus;
         if (interaction.guild) {
             // If user is on the list, and so is the guild that the interaction is called from, remove user and guild from list
             // Otherwise add user to list along with the guild that the interaction was called from (or just the guild if user is already on list)
-            if
-            (
+            if (
                 client.roles.roles.has(interaction.user.id) &&
                 client.roles.roles.get(interaction.user.id)?.has(interaction.guild.id)
-            )
-            {
+            ) {
                 client.roles.deleteUser(interaction.user, interaction.guild);
                 await interaction.reply("_OK, I won't randomise your colour any more._");
             } else {
